Fix: UT failures
